+++ grow/settings.py
@@ -5,13 +5,11 @@ from urllib.parse import urlparse
 
 import dj_database_url
 from the_music_tree_api_kit.utils.allowed_hosts import add_loopback_hosts
-from the_music_tree_genre_kit.data import DATA_DIR
 
 BASE_DIR = Path(__file__).resolve().parent.parent
 
 SECRET_KEY = os.environ["SECRET_KEY"]
 
-SYSTEM_USERNAME = os.environ["SYSTEM_USERNAME"]
 
 PIPELINE_API_KEY = os.environ["PIPELINE_API_KEY"]
 
@@ -90,7 +88,7 @@ REST_FRAMEWORK = {
         "djangorestframework_camel_case.parser.CamelCaseMultiPartParser",
         "djangorestframework_camel_case.parser.CamelCaseFormParser",
     ),
-    "DEFAULT_PERMISSION_CLASSES": ["rest_framework.permissions.IsAuthenticated"],
+    "DEFAULT_PERMISSION_CLASSES": ["grow.view.permission.IsAdminOrReadOnly.IsAdminOrReadOnly"],
     "EXCEPTION_HANDLER": "the_music_tree_api_kit.view.error.exception_handler.custom_exception_handler",
 }