+++ tests/integration/criteria/GenreTestCase.py
@@ -26,12 +26,6 @@ class GenreTestCase(AppTestCase):
             handle_response=self._set_results,
         )
 
-    def _post_genres_tree_load_seed(self):
-        return self.api_client.post(
-            path=reverse(self.list_endpoint) + "tree/load-seed/",
-            handle_response=self._set_results,
-        )
-
     def _post_genre(self, data=None):
         return self.api_client.post(
             path=reverse(self.list_endpoint), data=data, handle_response=self._set_error_response_result_if_failure
@@ -47,6 +41,12 @@ class GenreTestCase(AppTestCase):
     def _delete_genre(self, uuid):
         return self.api_client.delete(
             path=reverse("genre-detail", kwargs={"pk": uuid}),
+            handle_response=self._set_error_response_result_if_failure,
+        )
+
+    def _post_genre_exclude(self, uuid):
+        return self.api_client.post(
+            path=reverse("genre-detail", kwargs={"pk": uuid}) + "exclude/",
             handle_response=self._set_error_response_result_if_failure,
         )